Update, gets team names, trying to get other stats
The scraper now has the team name at a player row level, from here the the player information can beadded to the team list table and the match information can be created. More work required to get the other stats that are desirable such as the ground name, attendance date, time, etc.